+++ src/main/java/com/cts/capstone/service/SupplierService.java
@@ -2,10 +2,10 @@ package com.cts.capstone.service;
 
 import com.cts.capstone.model.Supplier;
 import com.cts.capstone.repository.SupplierRepository;
-import org.springframework.dao.EmptyResultDataAccessException;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
+import java.util.Optional;
 
 
 @Service
@@ -34,17 +34,13 @@ public class SupplierService {
 		return supplierRepository.save(product);
 	}
 
-	public void delete(Supplier supplier) {
-		try {
-			supplierRepository.delete(supplier);
-		} catch (IllegalArgumentException ignored) {
-		}
-	}
-
-	public void delete(Long id) {
-		try {
+	public boolean delete(Long id) {
+		Optional<Supplier> supplier = supplierRepository.findById(id);
+		if (supplier.isPresent()) {
 			supplierRepository.deleteById(id);
-		} catch (IllegalArgumentException | EmptyResultDataAccessException ignored) {
+			return true;
+		} else {
+			return false;
 		}
 	}
 }
